Commit to bi-furcated associations.
I have updated the User and Request models to take a bi-furcated approach
and added in methods to consolidate both sides of the relationship into
one single call.

This way, I can actually build custom methods to the User model to handle
things like adding friends, etc without much fuss.

The one caution I have is to be very clear about "friend" vs "request" so
that it is clear what is going on.